Changing Parent to Master Material for Materials, Adding New House/Door/Tree Models, Adding Textures for House/Door/Tree Models, Creating Materials for House/Door/TreeModels
I changed the Parents to the Master Material for all of my Materials and added the textures in the dropdown boxes for those materials. I added the New Houses and New Door models with the scales fixed and added the New Tree Models. I added the textures for the New Houses, New Doors and New Tree models. I created materials for the New Houses, New Doors and New Tree models.