Adding Models, Textures, Creating Materials and Changing Names of Materials
In this update I added my newest models which include NewTree4, NewTree5, NewTree6, Workbench and Radio. I added textures for all of those models and created materials for those assets ready to be placed on the models. I also changed the names of some of the materials that I created so that they look uniform with the rest. This is what makes it look like some of the materials were deleted here, but it was just the old names that were deleted, and new ones were created.