Finished Retextures
Finished uploading the retextures with the inclusion of the worn variants of the textures. I also saw a lot of odd duplicates of textures/materials I already uploaded; these duplicates had the same names as the original files, but their content were blank. I decided to delete them but if there was a reason for their inclusion please let me know and feel free to reverse it.